move logic to hook
